requirements.txt
streamlit==1.32.0
yfinance==0.2.37
pandas==2.2.1
numpy==1.26.4
plotly==5.20.0
scikit-learn==1.4.1
ta==0.11.0
requests==2.31.0
deap==1.4.1
tensorflow-cpu==2.15.0